requirements.txt
asgiref==3.5.0
Django==4.0.2
django-tastypie==0.14.4
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
